+++ NEWS.md
@@ -214,6 +214,8 @@
 
     - `lang()` has been renamed to `call2()`.
     - `new_language()` has ben renamed to `new_call()`.
+    - `is_lang()`, `is_unary_lang()`, and `is_binary_lang()` have been
+      renamed to `is_call()`, `is_unary_call()`, and `is_binary_call()`.
 
   In rlang 0.1 calls were called "language" objects in order to follow
   the R type nomenclature as returned by `base::typeof()`. We wanted